pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0"]

[tool.poetry]
name = "dotenv-linter"
version = "0.6.0"
description = "Linting dotenv files like a charm!"
authors = [
  "sobolevn <mail@sobolevn.me>"
]
license = "MIT"

readme = "README.md"

repository = "https://github.com/wemake-services/dotenv-linter"
homepage = "https://dotenv-linter.readthedocs.io"

keywords = [
  "dotenv",
  "linter",
  "wemake.services",
  "code quality"
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Quality Assurance",
]

[tool.poetry.dependencies]
python = "^3.10"

ply = "^3.11"
click = ">=6,<9"
click_default_group = "^1.2"
attrs = "*"
typing_extensions = ">=4.0,<5.0"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^1.0"
ruff = "^0.9"

pytest = "^8.1"
pytest-cov = "^6.0"
pytest-randomly = "^3.12"

codespell = "^2.2"

mypy = "^1.15"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^8.1"
sphinx-autodoc-typehints = "^3.0"
sphinx-click = "^6.0"
tomli = "^2.0"
myst-parser = "^4.0"
furo = "^2024.8"

[tool.poetry.scripts]
dotenv-linter = "dotenv_linter.cli:cli"


[tool.ruff]
# Ruff config: https://docs.astral.sh/ruff/settings
preview = true
fix = true
target-version = "py310"
line-length = 80
extend-exclude = [
  # Intentionally bad code:
  "dotenv_linter/grammar/parsetab.py",
]

[tool.ruff.format]
quote-style = "single"
docstring-code-format = false

[tool.ruff.lint]
select = [
  "A",    # flake8-builtins
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "C90",  # maccabe
  "COM",  # flake8-commas
  "D",    # pydocstyle
  "DTZ",  # flake8-datetimez
  "E",    # pycodestyle
  "ERA",  # flake8-eradicate
  "EXE",  # flake8-executable
  "F",    # pyflakes
  "FBT",  # flake8-boolean-trap
  "FLY",  # pyflint
  "FURB", # refurb
  "G",    # flake8-logging-format
  "I",    # isort
  "ICN",  # flake8-import-conventions
  "ISC",  # flake8-implicit-str-concat
  "LOG",  # flake8-logging
  "N",    # pep8-naming
  "PERF", # perflint
  "PIE",  # flake8-pie
  "PL",   # pylint
  "PT",   # flake8-pytest-style
  "PTH",  # flake8-use-pathlib
  "Q",    # flake8-quotes
  "RET",  # flake8-return
  "RSE",  # flake8-raise
  "RUF",  # ruff
  "S",    # flake8-bandit
  "SIM",  # flake8-simpify
  "SLF",  # flake8-self
  "SLOT", # flake8-slots
  "T100", # flake8-debugger
  "TRY",  # tryceratops
  "UP",   # pyupgrade
  "W",    # pycodestyle
  "YTT",  # flake8-2020
]
ignore = [
  "A005",   # allow to shadow stdlib and builtin module names
  "COM812", # trailing comma, conflicts with `ruff format`
  # Different doc rules that we don't really care about:
  "D100",
  "D104",
  "D106",
  "D203",
  "D212",
  "D401",
  "D404",
  "D405",
  "ISC001",  # implicit string concat conflicts with `ruff format`
  "ISC003",  # prefer explicit string concat over implicit concat
  "PLR09",   # we have our own complexity rules
  "PLR2004", # do not report magic numbers
  "PLR6301", # do not require classmethod / staticmethod when self not used
  "TRY003",  # long exception messages from `tryceratops`
]
external = [ "WPS" ]

# Plugin configs:
flake8-quotes.inline-quotes = "single"
mccabe.max-complexity = 6
pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"dotenv_linter/grammar/parser.py" = [
  "D205", # multiline grammar definitions
  "D403", # lowercase docstrings are fine
  "D415", # we cannot add `.` in the end, it is a grammar definition
]
"dotenv_linter/grammar/lexer.py" = [
  "N802", # lexer name patterns
]
"tests/*.py" = [
  "S101", # asserts
  "S404", # subprocess
  "S603", # subprocess
  "S607", # subprocess partial path
]
